Feature/trip co path client

+++ trip_company/service/company.go
@@ -4,6 +4,9 @@ import (
 	"context"
 	"errors"
 	"tripcompanyservice/internal/company"
+	"tripcompanyservice/internal/trip"
+
+	"github.com/google/uuid"
 )
 
 var (
@@ -12,11 +15,13 @@ var (
 
 type TransportCompanyService struct {
 	companyOps *company.Ops
+	tripOps    *trip.Ops
 }
 
-func NewTransportCompanyService(companyOps *company.Ops) *TransportCompanyService {
+func NewTransportCompanyService(companyOps *company.Ops, tripOps *trip.Ops) *TransportCompanyService {
 	return &TransportCompanyService{
 		companyOps: companyOps,
+		tripOps:    tripOps,
 	}
 }
 
@@ -24,7 +29,7 @@ func (s *TransportCompanyService) CreateTransportCompany(ctx context.Context, co
 	return s.companyOps.Create(ctx, company)
 }
 
-func (s *TransportCompanyService) GetUserTransportCompanies(ctx context.Context, ownerID uint, page, pageSize uint) ([]company.TransportCompany, uint, error) {
+func (s *TransportCompanyService) GetUserTransportCompanies(ctx context.Context, ownerID uuid.UUID, page, pageSize uint) ([]company.TransportCompany, uint, error) {
 	// user, err := s.userOps.GetUserByID(ctx, userID) check by the other service
 	// if user == nil {
 	// 	return nil, 0, u.ErrUserNotFound
@@ -56,10 +61,22 @@ func (s *TransportCompanyService) BlockCompany(ctx context.Context, companyID ui
 	return transportCompany, nil
 }
 
-func (s *TransportCompanyService) DeleteCompany(ctx context.Context, companyID uint) error {
-	_, err := s.companyOps.GetByID(ctx, companyID)
+func (s *TransportCompanyService) DeleteCompany(ctx context.Context, companyID uint, requesterID uuid.UUID) error {
+	co, err := s.companyOps.GetByID(ctx, companyID)
 	if err != nil {
 		return err
+	}
+
+	nActiveTrips, err := s.tripOps.GetCountCompanyUnfinishedUncanceledTrips(ctx, companyID)
+
+	if err!=nil{
+		return err
+	}
+	if nActiveTrips > 0 {
+		return company.ErrCanNotDelete
+	}
+	if co.OwnerID != requesterID{
+		return ErrForbidden
 	}
 	err = s.companyOps.Delete(ctx, companyID)
 	if err != nil {
@@ -68,7 +85,7 @@ func (s *TransportCompanyService) DeleteCompany(ctx context.Context, companyID u
 	return nil
 }
 
-func (s *TransportCompanyService) PatchCompanyByOwner(ctx context.Context, updatedCompany *company.TransportCompany, userID uint, newOwnerEmail string) (*company.TransportCompany, error) {
+func (s *TransportCompanyService) PatchCompanyByOwner(ctx context.Context, updatedCompany *company.TransportCompany, userID uuid.UUID, newOwnerEmail string) (*company.TransportCompany, error) {
 	originalCompany, err := s.companyOps.GetByID(ctx, updatedCompany.ID)
 	if err != nil {
 		return nil, err
@@ -77,11 +94,11 @@ func (s *TransportCompanyService) PatchCompanyByOwner(ctx context.Context, updat
 		return nil, ErrForbidden
 	}
 
-	if newOwnerEmail!=""{
+	if newOwnerEmail != "" {
 		//******
 		// TO DO: check this user exists and get it!!!!
-		updatedCompany.OwnerID = 1
-		return nil, nil
+		//updatedCompany.OwnerID = 1
+		return nil, errors.New("we need to find user's uuid to withdraw the company")
 	}
 
 	err = s.companyOps.PatchCompanyByOwner(ctx, updatedCompany, originalCompany)
